insert for b tree

+++ dbms/src/orange/table/column.h
@@ -69,7 +69,7 @@ public:
     bool is_indexed() { return indexed; }
     index_key_kind_t key_kind() {
         switch (datatype.kind) {
-            case datatype_t::NUMERIC: return index_key_kind_t::FLOAT;
+            case datatype_t::NUMERIC: return index_key_kind_t::NUMERIC;
             case datatype_t::VARCHAR: return index_key_kind_t::VARCHAR;
             default: return index_key_kind_t::BYTES;
         }
